+++ src/main/java/org/prgrms/springbootbasic/repository/voucher/MemoryVoucherRepository.java
@@ -1,4 +1,4 @@
-package org.prgrms.springbootbasic.repository;
+package org.prgrms.springbootbasic.repository.voucher;
 
 import java.util.ArrayList;
 import java.util.List;
