+++ src/test/java/org/prgrms/springbootbasic/repository/MemoryVoucherRepositoryTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.prgrms.springbootbasic.entity.FixedAmountVoucher;
 import org.prgrms.springbootbasic.entity.PercentDiscountVoucher;
 import org.prgrms.springbootbasic.entity.Voucher;
+import org.prgrms.springbootbasic.repository.voucher.MemoryVoucherRepository;
 
 class MemoryVoucherRepositoryTest {
