+++ src/test/java/org/prgrms/springbootbasic/service/VoucherServiceTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.prgrms.springbootbasic.repository.MemoryVoucherRepository;
+import org.prgrms.springbootbasic.repository.voucher.MemoryVoucherRepository;
 
 class VoucherServiceTest {
